pyproject.toml
[project]
name = "rubish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "kurigram>=2.2.19",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "tgcrypto>=1.2.5",
    "uvloop>=0.22.1",
]
